% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sr_reflectance_data}
\alias{sr_reflectance_data}
\title{SUNGLINT Removal (SR) for WISP station reflectance data (Jiang et al. (2020))}
\usage{
sr_reflectance_data(qc_data)
}
\arguments{
\item{qc_data}{A \code{tibble} from qc_reflectance_data() function.}
}
\description{
A short description...

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# example code
\dontrun{
## Not run:
reflec_data_sr <- WISP.data::sr_reflectance_data(qc_data = reflec_data_qc) 
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
