% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_get_reflectance_data}
\alias{wisp_get_reflectance_data}
\title{Get data of reflectance (level2) from WISPstation for a specific date}
\usage{
wisp_get_reflectance_data(
  version = "1.0",
  time_from = NULL,
  time_to = NULL,
  station = NULL,
  userid = NULL,
  pwd = NULL,
  save_csv = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{version}{A \code{character}. It is the version of the API data. Default
is "1.0".}

\item{time_from}{A \code{character}. It is the date and time from which the data
is requested.}

\item{time_to}{A \code{character}. It is the date and time to which the data
is requested.}

\item{station}{A \code{character}. It is the name of the station.}

\item{userid}{A \code{character}. It is the userid to access to the data service.}

\item{pwd}{A \code{character}. It is the password to access to the data service.}

\item{save_csv}{A \code{logical}. If \code{TRUE}, the function saves the reflectance data.}

\item{out_dir}{A \code{character}. The directory where the CSV file will be saved.
Default is \code{NULL}, which saves to a temporary directory (\code{tempdir()}) if \code{save_csv = TRUE}.}
}
\value{
A \code{tibble} with measurement id, measurement date, instrument name,
level2_quality, set of sensor (irradiance and radiances),
water quality values of TSM, Chla, Kd, and cpc as provided by instrument by default,
all the reflectance values from 350 to 900 nm.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function represents the main entry point for data acquisition within the
WISP.data package. It connects to the official Water Insight APIs to download
spectral reflectance measurements for a user-defined time interval.
The function handles authentication queries the remote server and returns
the retrieved data in a structured tibble format. In addition to hyperspectral
reflectance data (350-900 nm), the function also retrieves the water quality
parameters natively computed by the WISPstation, including: TSM
(Van Der Woerd & Pasterkamp, 2008), Chla (Gons et al., 2005), Kd
(Gons et al., 1998), and cpc (Simis, 2006).
}
\examples{
\dontrun{
# NA data
reflect_data <- wisp_get_reflectance_data(
  time_from = "2024-09-01T09:00",
  time_to = "2024-09-01T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd,
  save_csv = FALSE
)

# with data
reflect_data <- wisp_get_reflectance_data(
  time_from = "2024-08-01T09:00",
  time_to = "2024-08-01T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd,
  save_csv = FALSE
)

# no data for the station selected
reflect_data <- wisp_get_reflectance_data(
  time_from = "2019-06-20T09:00",
  time_to = "2019-06-20T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd,
  save_csv = FALSE
)

# The two dates are not consistent
reflect_data <- wisp_get_reflectance_data(
  time_from = "2019-06-20T09:00",
  time_to = "2020-06-20T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd,
  save_csv = FALSE
)
}

}
\author{
Alessandro Oggioni, phD \href{mailto:alessandro.oggioni@cnr.it}{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \href{mailto:nicola.ghirardi@cnr.it}{nicola.ghirardi@cnr.it}
}
