% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_novoa_SPM}
\alias{wisp_novoa_SPM}
\title{Calculation of SPM concentration (Novoa et al., 2017)}
\usage{
wisp_novoa_SPM(qc_data, sr_data, calc_SPM = TRUE)
}
\arguments{
\item{qc_data}{A \code{tibble} from wisp_qc_reflectance_data() function.}

\item{sr_data}{A \code{tibble} from wisp_sr_reflectance_data() function.}

\item{calc_SPM}{A \code{logical}. If \code{TRUE}, the function calculates the
NOVOA SPM concentrations. Default is \code{TRUE}.}
}
\value{
The result is a list with two elements: ...
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the SPM concentration in according
to Novoa et al., 2017 and creates new columns in qc_data e sr_data.
\code{calc_SPM} can be either TRUE (default) or FALSE.
}
\examples{
# example code
\dontrun{
#' ## Not run:

# Calculate NOVOA_SPM
novoa_SPM_results <- WISP.data::wisp_novoa_SPM(qc_data = reflect_data_qc, sr_data = reflect_data_sr, calc_SPM = T)
reflect_data_qc <- novoa_SPM_results$qc_data
reflect_data_sr <- novoa_SPM_results$sr_data

# Don't calculate NOVOA_SPM
novoa_SPM_results <- WISP.data::wisp_novoa_SPM(qc_data = reflect_data_qc, sr_data = reflect_data_sr, calc_SPM = F)
reflect_data_qc <- novoa_SPM_results$qc_data
reflect_data_sr <- novoa_SPM_results$sr_data
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
