% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_novoa_TUR}
\alias{wisp_novoa_TUR}
\title{Calculation of Turbidity (FNU) (Novoa et al., 2017)}
\usage{
wisp_novoa_TUR(qc_data, sr_data, calc_TUR = TRUE)
}
\arguments{
\item{qc_data}{A \code{tibble} from wisp_qc_reflectance_data() function.}

\item{sr_data}{A \code{tibble} from wisp_sr_reflectance_data() function.}

\item{calc_TUR}{A \code{logical}. If \code{TRUE}, the function calculates the
NOVOA TUR values Default is \code{TRUE}.}
}
\value{
The result is a list with two elements: ...
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the Turbidity (FNU) in according
to Novoa et al., 2017 and creates new columns in qc_data e sr_data.
\code{calc_TUR} can be either TRUE (default) or FALSE.
}
\examples{
# example code
\dontrun{
## Not run:

# Calculate NOVOA_SPM
novoa_TUR_results <- WISP.data::wisp_novoa_TUR(qc_data = reflect_data_qc, sr_data = reflect_data_sr, calc_TUR = T)
reflect_data_qc <- novoa_TUR_results$qc_data
reflect_data_sr <- novoa_TUR_results$sr_data

# Don't calculate NOVOA_SPM
novoa_TUR_results <- WISP.data::wisp_novoa_TUR(qc_data = reflect_data_qc, sr_data = reflect_data_sr, calc_TUR = F)
reflect_data_qc <- novoa_TUR_results$qc_data
reflect_data_sr <- novoa_TUR_results$sr_data
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
