% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_phyto_scattering}
\alias{wisp_phyto_scattering}
\title{Phytoplankton Scattering for WISPstation reflectance data}
\usage{
wisp_phyto_scattering(qc_data, sr_data, calc_scatt = TRUE)
}
\arguments{
\item{qc_data}{A \code{tibble} from wisp_qc_reflectance_data() function.}

\item{sr_data}{A \code{tibble} from wisp_sr_reflectance_data() function.}

\item{calc_scatt}{A \code{logical}. If \code{TRUE}, the function calculates the
two parameters. Default is \code{TRUE}.}
}
\value{
The result is a list with two elements: ...
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the reflectance peak due to phytoplankton
scattering (690-710 nm) and calculates the ratio of the latter
to the second chlorophyll absorption peak (670-680 nm).
`calc_scatt``` can be either TRUE (default) or FALSE.
}
\examples{
# example code
\dontrun{
## Not run:

# Calculate the two parameters
scattering_results <- WISP.data::wisp_phyto_scattering(qc_data = reflect_data_qc, sr_data = reflect_data_sr, calc_scatt = T)
reflect_data_qc <- scattering_results$qc_data
reflect_data_sr <- scattering_results$sr_data

# Don't calculate the two parameters
scattering_results <- WISP.data::wisp_phyto_scattering(qc_data = reflect_data_qc, sr_data = reflect_data_sr, calc_scatt = F)
reflect_data_qc <- scattering_results$qc_data
reflect_data_sr <- scattering_results$sr_data

}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
