% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_plot_reflectance_data}
\alias{wisp_plot_reflectance_data}
\title{Create a plot of reflectance data}
\usage{
wisp_plot_reflectance_data(
  data,
  legend_TSM = TRUE,
  legend_Chla = TRUE,
  legend_Kd = TRUE,
  legend_cpc = TRUE,
  legend_scatt = FALSE,
  legend_ratio = FALSE,
  legend_novoa_SPM = FALSE,
  legend_novoa_TUR = FALSE,
  legend_jiang_TSS = FALSE,
  legend_gons_CHL = FALSE,
  legend_gons740_CHL = FALSE,
  legend_NDCI = FALSE,
  legend_mishra_CHL = FALSE,
  legend_hue_angle = FALSE,
  legend_dom_wavelength = FALSE,
  legend_OWT_class = FALSE,
  legend_OWT_score = FALSE,
  legend_OWT_z_dist = FALSE
)
}
\arguments{
\item{data}{A \code{tibble} obtained by any of the functions provided by this
package: \code{wisp_get_reflectance_data()}, or after QC and SR removal operations.}

\item{legend_TSM}{A \code{logical}. If \code{TRUE}, the plot legend includes the \code{TSM}
values. Default is \code{TRUE}.}

\item{legend_Chla}{A \code{logical}. If \code{TRUE}, the plot legend includes the
\code{Chla} values. Default is \code{TRUE}.}

\item{legend_Kd}{A \code{logical}. If \code{TRUE}, the plot legend includes the \code{Kd}
values. Default is \code{TRUE}.}

\item{legend_cpc}{A \code{logical}. If \code{TRUE}, the plot legend includes the \code{cpc}
values. Default is \code{TRUE}.}

\item{legend_scatt}{A \code{logical}. If \code{TRUE}, the plot legend includes the \code{scattering}
values. Default is \code{FALSE}.}

\item{legend_ratio}{A \code{logical}. If \code{TRUE}, the plot legend includes the \code{ratio}
values. Default is \code{FALSE}.}

\item{legend_novoa_SPM}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Novoa_SPM} values. Default is \code{FALSE}.}

\item{legend_novoa_TUR}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Novoa_TUR} values. Default is \code{FALSE}.}

\item{legend_jiang_TSS}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Jiang_TSS} values. Default is \code{FALSE}.}

\item{legend_gons_CHL}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Gons_CHL} values. Default is \code{FALSE}.}

\item{legend_gons740_CHL}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Gons740_CHL} values. Default is \code{FALSE}.}

\item{legend_NDCI}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{NDCI} values. Default is \code{FALSE}.}

\item{legend_mishra_CHL}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Mishra_CHL} values. Default is \code{FALSE}.}

\item{legend_hue_angle}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Hue_Angle} values. Default is \code{FALSE}.}

\item{legend_dom_wavelength}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{Dominant_Wavelength} values. Default is \code{FALSE}.}

\item{legend_OWT_class}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{OWT_class}. Default is \code{FALSE}.}

\item{legend_OWT_score}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{OWT_score} (membership grade). Default is \code{FALSE}.}

\item{legend_OWT_z_dist}{A \code{logical}. If \code{TRUE}, the plot legend includes
the \code{OWT_z_dist} (statistical distance). Default is \code{FALSE}.}
}
\value{
An interactive \code{plotly} object showing the spectral signatures of the
reflectance data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function generates an interactive visualization of all spectral signatures
contained in a dataset, based on the plotly library. It is highly flexible and
can be used to display: native data downloaded directly from WISPstation, processed
data after QC, processed data after SR. The function's distinctive feature is
its dynamic tooltip system: when hovering over a spectral curve, users can
instantly visualize the corresponding acquisition date and time, together with
all associated bio-optical parameters computed for that specific measurement.
}
\examples{
if (interactive()) {
  wisp_plot_reflectance_data(
    data = reflect_data_sr,
    legend_TSM = TRUE,
    legend_Chla = TRUE,
    legend_Kd = TRUE,
    legend_cpc = TRUE,
    legend_scatt = FALSE,
    legend_ratio = FALSE,
    legend_novoa_SPM = FALSE,
    legend_novoa_TUR = FALSE,
    legend_jiang_TSS = FALSE,
    legend_gons_CHL = FALSE,
    legend_gons740_CHL = FALSE,
    legend_NDCI = FALSE,
    legend_mishra_CHL = FALSE,
    legend_hue_angle = FALSE,
    legend_dom_wavelength = FALSE,
    legend_OWT_class = FALSE,
    legend_OWT_score = FALSE,
    legend_OWT_z_dist = FALSE
  )
}

}
\author{
Alessandro Oggioni, phD \href{mailto:alessandro.oggioni@cnr.it}{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \href{mailto:nicola.ghirardi@cnr.it}{nicola.ghirardi@cnr.it}
}
