% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisp_runApp.R
\name{wisp_runApp}
\alias{wisp_runApp}
\title{Run shiny app for get and visualize WISP data}
\usage{
wisp_runApp(stations = c("WISPstation012", "WISPstation013"), ...)
}
\arguments{
\item{stations}{A \verb{character vector} of station names.}

\item{...}{Other parameters passed to \code{shiny::runApp} (e.g. \code{launch.browser},
\code{port}, \code{host}).}
}
\value{
No return value, called for side effects to launch the interactive Shiny application.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function runs the Shiny app for querying and visualizing data from a specific WISP station.
}
\examples{
if (interactive()) {
  # Launch the Shiny application
  wisp_runApp(launch.browser = TRUE)
}
}
\author{
Alessandro Oggioni, phD \href{mailto:alessandro.oggioni@cnr.it}{alessandro.oggioni@cnr.it}
}
\keyword{internal}
