% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_sr_reflectance_data}
\alias{wisp_sr_reflectance_data}
\title{Sky-glint Removal (SR) for WISPstation reflectance data}
\usage{
wisp_sr_reflectance_data(
  qc_data,
  calc_scatt = TRUE,
  calc_SPM = TRUE,
  calc_TUR = TRUE,
  calc_TSS = TRUE,
  calc_gons = TRUE,
  calc_gons740 = TRUE,
  calc_NDCI = TRUE,
  calc_mishra = TRUE,
  calc_dom_wave = TRUE,
  calc_OWT = TRUE,
  save_csv = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{qc_data}{A \code{tibble} from \code{wisp_qc_reflectance_data()} function.}

\item{calc_scatt}{A \code{logical}. If \code{TRUE}, the function calculates the
peak due to phytoplankton scattering (690-710 nm) and the ratio of the latter
to the second chlorophyll absorption peak (670-680 nm). Default is \code{TRUE}.}

\item{calc_SPM}{A \code{logical}. If \code{TRUE}, the function calculates the
SPM concentrations in according to Novoa et al. (2017). Default is \code{TRUE}.}

\item{calc_TUR}{A \code{logical}. If \code{TRUE}, the function calculates the
turbidity (FNU) in according to Novoa et al. (2017). Default is \code{TRUE}.}

\item{calc_TSS}{A \code{logical}. If \code{TRUE}, the function calculates the
TSS concentrations in according to Jiang et al. (2021). Default is \code{TRUE}.}

\item{calc_gons}{A \code{logical}. If \code{TRUE}, the function calculates chlorophyll
using Gons et al. (2002) algorithm (NIR ~782 nm). Default is \code{TRUE}.}

\item{calc_gons740}{A \code{logical}. If \code{TRUE}, the function calculates chlorophyll
using Gons et al. (2002) algorithm (NIR ~740 nm). Default is \code{TRUE}.}

\item{calc_NDCI}{A \code{logical}. If \code{TRUE}, the function calculates The Normalised
Difference Chlorophyll Index algorithm by Mishra and Mishra (2012). Default is \code{TRUE}.}

\item{calc_mishra}{A \code{logical}. If \code{TRUE}, the function calculates chlorophyll
using Mishra and Mishra (2012) algorithm. Default is \code{TRUE}.}

\item{calc_dom_wave}{A \code{logical}. If \code{TRUE}, the function calculates the hue
angle and the dominant wavelength (van der Woerd and Wernand, 2015, 2018).
Default is \code{TRUE}.}

\item{calc_OWT}{A \code{logical}. If \code{TRUE}, the function calculates the Optical
Water Type classification in according to Bi and Hieronymi (2024) (OWT_class),
the membership probability (OWT_score), and the average Z-score
(OWT_z_dist). In addition, it adds a column with the description of the
corresponding OWT class. "OWT_class" represents the optical category to which
the analyzed spectral signature belongs based on Gaussian Likelihood.
"OWT_score" is a value between 0 and 1 representing the fuzzy membership grade;
it indicates the probability of belonging to the selected class relative to
the other available classes. "OWT_z_dist" indicates the statistical
distance between the observed spectrum and the class mean, weighted by its
standard deviation. 0-1.5 indicate an excellent fit with the OWT reference.
Values above 3 suggest that the spectral signature is an outlier or deviates
significantly from the typical range of that class. Default is \code{TRUE}.}

\item{save_csv}{A \code{logical}. If \code{TRUE}, the function saves the reflectance data.}

\item{out_dir}{A \code{character}. The directory where the CSV file will be saved.
Default is \code{NULL}, which saves to a temporary directory (\code{tempdir()}) if \code{save_csv = TRUE}.}
}
\value{
A \code{tibble} with the spectral signatures after the SR operation
and all the extra parameters that were requested. If parameter save_csv is
\code{TRUE}, the function saves the reflectance data in a CSV file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function implements a sky-glint Removal algorithm based on the methodology
proposed by Jiang et al. (2020). The function operates on spectra that have
already been filtered through the Quality Control process and applies a
correction to obtain glint-corrected remote sensing reflectance.
Following this correction, all algorithms are re-applied using the corrected
reflectance, ensuring more accurate and physically consistent estimates
of water constituents.
}
\examples{
\donttest{
# Requires a valid dataset output from wisp_qc_reflectance_data()
if (exists("reflect_data_qc")) {
  reflect_data_sr <- wisp_sr_reflectance_data(
    qc_data = reflect_data_qc,
    calc_scatt = TRUE,
    calc_SPM = TRUE,
    calc_TUR = TRUE,
    calc_TSS = TRUE,
    calc_gons = TRUE,
    calc_gons740 = TRUE,
    calc_NDCI = TRUE,
    calc_mishra = FALSE,
    calc_dom_wave = TRUE,
    calc_OWT = TRUE,
    save_csv = FALSE,
  )
}
}

}
\author{
Alessandro Oggioni, phD \href{mailto:alessandro.oggioni@cnr.it}{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \href{mailto:nicola.ghirardi@cnr.it}{nicola.ghirardi@cnr.it}
}
