% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qc_reflectance_data}
\alias{qc_reflectance_data}
\title{Quality check (QC) for WISP station reflectance data}
\usage{
qc_reflectance_data(data)
}
\arguments{
\item{data}{A \code{tibble}. From wisp_get_reflectance_data() function.}
}
\description{
A short description...

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# example code
WISP.data::qc_reflectance_data(data = reflec_data)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
