% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WISP.data-package.R
\docType{package}
\name{WISP.data-package}
\alias{WISP.data}
\alias{WISP.data-package}
\title{WISP.data: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oggioniale/WISP.data}
  \item Report bugs at \url{https://github.com/oggioniale/WISP.data/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Oggioni \email{alessandro.oggioni@cnr.it} (\href{https://orcid.org/0000-0002-7997-219X}{ORCID}) [funder]

Authors:
\itemize{
  \item Nicola Ghirardi \email{nicola.ghirardi@cnr.it} (\href{https://orcid.org/0000-0002-5006-9853}{ORCID})
}

}
\keyword{internal}
