% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_get_reflectance_data}
\alias{wisp_get_reflectance_data}
\title{Get data of reflectance (level2) from WISPstation for a specific date}
\usage{
wisp_get_reflectance_data(
  version = "1.0",
  time_from = NULL,
  time_to = NULL,
  station = NULL,
  userid = NULL,
  pwd = NULL
)
}
\arguments{
\item{version}{A \code{character}. It is the version of the API data. Default
is "1.0"}

\item{time_from}{A \code{character}. It is the date and time from which the data
is requested.}

\item{time_to}{A \code{character}. It is the date and time to which the data
is requested.}

\item{station}{A \code{character}.  It is the name of the station.}

\item{userid}{A \code{character}. It is the userid to access to the data service.}

\item{pwd}{A \code{character}. It is the password to access to the data service.}
}
\value{
A \code{tibble} with measurement id, measurement date, instrument name,
level2_quality, set of sensor (irradiance and radiances),
waterquality values of TSM, Chla, Kd, and cpc as provided by instrument by default,
all the values of reflectance for each wevelength from 350 to 900 nm.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function obtains the data of reflectance from WISPstation for a
specific date.
}
\examples{
# example code
\dontrun{
## Not run:
# NA data
reflect_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2024-09-01T09:00",
  time_to = "2024-09-01T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)

# with data
reflect_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2024-08-01T09:00",
  time_to = "2024-08-01T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)

# no data for the station selected
reflect_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2019-06-20T09:00",
  time_to = "2019-06-20T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)

# The two dates are not consistent
reflect_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2019-06-20T09:00",
  time_to = "2020-06-20T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
