% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_get_reflectance_data}
\alias{wisp_get_reflectance_data}
\title{Get data of reflectance (level2) from WISP station}
\usage{
wisp_get_reflectance_data(
  version = "1.0",
  time_from = NULL,
  time_to = NULL,
  station = NULL,
  userid = NULL,
  pwd = NULL
)
}
\arguments{
\item{version}{A \code{character}. It is the version of the API data. Default
is "1.0"}

\item{time_from}{A \code{character}. It is the time}

\item{time_to}{A \code{character}. It is the time}

\item{station}{A \code{character}.  It is the name of the station.}

\item{userid}{A \code{character}. It is the userid.}

\item{pwd}{A \code{character}. It is the password.}
}
\value{
A tibble with measurement id, measurement date, instrument name,
waterquality values of TSM, Chla and Kd as provided by instrument by default,
and all the values of reflectance for each wevelength from 350 to 900 nm.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function obtains the data of reflectance from WISP station.
}
\examples{
# example code
\dontrun{
## Not run:
# NA data
reflec_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2024-09-01T09:00",
  time_to = "2024-09-01T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)

# with data
reflec_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2024-08-01T09:00",
  time_to = "2024-08-01T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)

# no data for the station selected
reflec_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2019-06-20T09:00",
  time_to = "2019-06-20T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)

# The two dates are not consistent
reflec_data <- WISP.data::wisp_get_reflectance_data(
  time_from = "2019-06-20T09:00",
  time_to = "2020-06-20T14:00",
  station = "WISPstation012",
  userid = userid,
  pwd = pwd
)
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}
}
