% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_plot_comparison}
\alias{wisp_plot_comparison}
\title{Comparison plot of Raw vs QC vs SR reflectance data}
\usage{
wisp_plot_comparison(
  raw_data,
  qc_data = NULL,
  sr_data = NULL,
  raw_args = NULL,
  qc_args = NULL,
  sr_args = NULL
)
}
\arguments{
\item{raw_data}{A \code{tibble}. The original data obtained by \code{wisp_get_reflectance_data()}.}

\item{qc_data}{A \code{tibble}. The data after \code{wisp_qc_reflectance_data()} operations.
Default is \code{NULL}.}

\item{sr_data}{A \code{tibble}. The data after \code{wisp_sr_reflectance_data()} operations.
Default is \code{NULL}.}

\item{raw_args}{A \code{list} of arguments to be passed to \code{wisp_plot_reflectance_data}
for the raw data plot (legend). Default is \code{NULL}.}

\item{qc_args}{A \code{list} of arguments to be passed to \code{wisp_plot_reflectance_data}
for the QC data plot (legend). Default is \code{NULL}.}

\item{sr_args}{A \code{list} of arguments to be passed to \code{wisp_plot_reflectance_data}
for the SR data plot (legend). Default is \code{NULL}.}
}
\value{
A \code{plotly} subplot object comparing the spectral signatures (Raw vs QC
vs SR). If only \code{raw_data} is provided or valid, a single plot is returned.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function creates an interactive side-by-side visual comparison of the different
WISPstation data processing levels. Using plotly submodules, it enables the
visualization of up to three aligned plots within a single interactive window:
native data downloaded directly from WISPstation, processed data after QC,
processed data after SR. This provides a powerful tool for visually assessing
how filtering and correction algorithms modify spectral signatures, remove
artifacts, and improve data quality.
}
\examples{
# example code
\dontrun{
## Not run:
custom_raw <- list(legend_TSM = FALSE, legend_Chla = FALSE)
custom_qc  <- list(legend_TSM = TRUE, legend_Chla = TRUE, legend_Kd = FALSE)
custom_sr  <- list(legend_TSM = TRUE, legend_mishra_CHL = FALSE)

fig_comparison <- wisp_plot_comparison(
  raw_data = reflect_data,
  qc_data  = reflect_data_qc,
  sr_data  = reflect_data_sr,
  raw_args = custom_raw,
  qc_args  = custom_qc,
  sr_args  = custom_sr
)
print(fig_comparison)
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
