% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_qc_reflectance_data}
\alias{wisp_qc_reflectance_data}
\title{Quality Control (QC) for WISPstation reflectance data}
\usage{
wisp_qc_reflectance_data(
  data,
  maxPeak = 0.05,
  maxPeak_blue = 0.02,
  qa_threshold = 0.5,
  qwip_threshold = 0.2,
  calc_scatt = TRUE,
  calc_SPM = TRUE,
  calc_TUR = TRUE,
  calc_TSS = TRUE,
  save_csv = FALSE,
  out_dir = "outputs"
)
}
\arguments{
\item{data}{A \code{tibble}. From wisp_get_reflectance_data() function.}

\item{maxPeak}{A \code{decimal}. Maximum magnitude of the spectral signatures.
We recommend setting this parameter to: 0.02 for clear and oligotrophic water,
0.05 for meso- to eutrophic water, and 0.08 for hypereutrophic and highly turbid water.
Default is 0.05.}

\item{maxPeak_blue}{A \code{decimal}. Maximum magnitude 350 nm values.
We recommend setting this parameter to: 0.02 (default)}

\item{qa_threshold}{A \code{decimal}. Minimum threshold for Quality Assurance (QA).
We recommend setting this parameter to: 0.5 (default)}

\item{qwip_threshold}{A \code{decimal}. Maximum threshold for Quality Water Index Polynomial (QWIP).
We recommend setting this parameter to: 0.2 (default)}

\item{calc_scatt}{A \code{logical}. If \code{TRUE}, the function calculates the
peak due to phytoplankton scattering (690-710 nm) and the ratio of the latter
to the second chlorophyll absorption peak (670-680 nm). Default is \code{TRUE}.}

\item{calc_SPM}{A \code{logical}. If \code{TRUE}, the function calculates the
SPM concentrations in according to Novoa et al., (2017). Default is \code{TRUE}.}

\item{calc_TUR}{A \code{logical}. If \code{TRUE}, the function calculates the
turbidity (FNU) in according to Novoa et al., (2017). Default is \code{TRUE}.}

\item{calc_TSS}{A \code{logical}. If \code{TRUE}, the function calculates the
TSS concentrations in according to Jiang et al., (2021). Default is \code{TRUE}.}

\item{save_csv}{A \code{logical}. If \code{TRUE}, the function saves the reflectance data.}

\item{out_dir}{A \code{character}. The directory where the CSV file will be saved.
Default is "outputs" within the working directory.}
}
\value{
A \code{tibble} with the spectral signatures that have passed QC operation. In addition,
a message containing the reason behind the elimination of each anomalous spectral signature
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function removes all anomalous spectral signatures and
explains the reason for each elimination
}
\examples{
# example code
\dontrun{
## Not run:
reflect_data_qc <- wisp_qc_reflectance_data(
  data = reflect_data,
  maxPeak = 0.05,
  maxPeak_blue = 0.02,
  qa_threshold    = 0.5,
  qwip_threshold  = 0.2,
  calc_scatt = TRUE,
  calc_SPM = FALSE,
  calc_TUR = TRUE,
  calc_TSS = TRUE,
  save_csv = FALSE,
  out_dir = "outputs"
  QA_ref_csv = "path/QA_raw_data_nRrs_Wei2016.csv"
)
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
