% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisp_runApp.R
\name{wisp_runApp}
\alias{wisp_runApp}
\title{Run shiny app for get and visualize WISP data}
\usage{
wisp_runApp(...)
}
\arguments{
\item{...}{other params passed to \code{shiny::runApp}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function run the Shiny App
}
\examples{
# example code
WISP.data::wisp_runApp(launch.browser=rstudioapi::viewer)
}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}
}
