% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wisp_sr_reflectance_data}
\alias{wisp_sr_reflectance_data}
\title{SUNGLINT Removal (SR) for WISPstation reflectance data}
\usage{
wisp_sr_reflectance_data(qc_data, save_csv = FALSE, out_dir = "outputs")
}
\arguments{
\item{qc_data}{A \code{tibble} from wisp_qc_reflectance_data() function.}

\item{save_csv}{A \code{logical}. If \code{TRUE}, the function saves the reflectance data.}

\item{out_dir}{A \code{character}. The directory where the CSV file will be saved.
Default is "outputs" within the working directory.}
}
\value{
A tibble with the spectral signatures after the SR operation and,
if parameter \code{save_out_sr} is \code{TRUE}, the function saves the reflectance data
in a CSV file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function applies the algorithm of Jiang et al., (2020) for removing
sunglint from spectral signatures
}
\examples{
# example code
\dontrun{
## Not run:
reflect_data_sr <- WISP.data::wisp_sr_reflectance_data(
  qc_data = reflect_data_qc,
  save_csv = TRUE
) 
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}

Nicola Ghirardi, phD \email{nicola.ghirardi@cnr.it}
}
